+++ src/models/channel/error.rs
@@ -3,7 +3,7 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum ChannelError {
     #[error("Requested channel was not found in channel model layer: {0}")]
-    NotFound(i8),
+    NotFound(i16),
 
     #[error("Unexpected error in channel model layer")]
     UnexpectedError,
