+++ src/net/packet/handler/check_char_name/builder.rs
@@ -14,7 +14,7 @@ impl Packet {
         let op = SendOpcode::CharNameResponse as i16;
         self.write_short(op).map_err(WriteError)?;
         self.write_str(ign).map_err(WriteError)?;
-        let exists = exists as i8;
+        let exists = exists as i16;
         self.write_byte(exists).map_err(WriteError)?;
         Ok(self)
     }
