+++ src/net/packet/handler/credentials/builder.rs
@@ -10,7 +10,7 @@ use std::time::UNIX_EPOCH;
 impl Packet {
     pub fn build_credentials_handler_failed_login_packet(
         &mut self,
-        status: i8,
+        status: i16,
     ) -> Result<&mut Self, NetworkError> {
         let op = SendOpcode::AccountStatus as i16;
         self.write_short(op).map_err(WriteError)?;
@@ -24,10 +24,10 @@ impl Packet {
         &mut self,
         acc_model: &AccountModel,
     ) -> Result<&mut Self, NetworkError> {
-        let pin_required = settings::get_pin_required()? as i8;
+        let pin_required = settings::get_pin_required()? as i16;
         let opcode = SendOpcode::AccountStatus as i16;
         let acc_id = acc_model.id as i32;
-        let gender_id = acc_model.gender_id as i8;
+        let gender_id = acc_model.gender_id as i16;
         let account_name = acc_model.username.clone();
         let created_at: i64 = acc_model
             .created_at
