+++ src/net/packet/handler/player_logged_in/reader.rs
@@ -7,7 +7,7 @@ use std::io::Cursor;
 #[derive(Clone)]
 pub struct PlayerLoggedInReader {
     pub char_id: i32,
-    pub channel_id: i8,
+    pub channel_id: i16,
 }
 
 impl PlayerLoggedInReader {
@@ -15,7 +15,7 @@ impl PlayerLoggedInReader {
         let mut pkt_reader = Cursor::new(&packet.bytes);
         let _op = pkt_reader.read_short().map_err(ReadError)?;
         let char_id = pkt_reader.read_int().map_err(ReadError)?;
-        let channel_id = pkt_reader.read_byte().map_err(ReadError)?;
+        let channel_id = pkt_reader.read_byte().map_err(ReadError)? as i16;
         Ok(Self {
             char_id,
             channel_id,
