+++ src/net/packet/handler/tos/reader.rs
@@ -6,14 +6,14 @@ use std::io::Cursor;
 
 #[derive(Clone)]
 pub struct TosReader {
-    pub confirmed: i8,
+    pub confirmed: i16,
 }
 
 impl TosReader {
     pub fn read_tos_packet(packet: &Packet) -> Result<Self, NetworkError> {
         let mut pkt_reader = Cursor::new(&packet.bytes);
         let _op = pkt_reader.read_short().map_err(ReadError)?;
-        let confirmed = pkt_reader.read_byte().map_err(ReadError)?;
+        let confirmed = pkt_reader.read_byte().map_err(ReadError)? as i16;
         Ok(Self {
             confirmed: confirmed,
         })
